Add verification logic for webhook token custom claims.
This change adds logic to extract and validate the custom message-digest claims that we are starting to embed in the tokens accompanying our webhook requests to further improve the security properties of Chainguard webhooks.

Now these authorization tokens effective sign the message payload they are intended to accompany, which makes it even less practical to spoof events as coming from Chainguard.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>